requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.1
Django==5.1
django-crispy-forms==2.3
django-mathfilters==1.0.0
djangorestframework==3.15.2
mysql-connector-python==9.0.0
mysqlclient==2.2.4
PyMySQL==1.1.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
